Rtcl 1.1 update release adds -verbose switch to all Rtcl commands.
This update adds optional -verbose switch to all Rtcl
commands. Presence of -verbose causes embedded R interpreter to print
the results of R expressions to stdout as they are evaluated. This may
be used to debug the results of R expressons as Rtcl commands are run
(see Rtcl man page). This was the default in 1.0 however R's stdout
messages are not used by Rtcl and caused excessive output to stdout.

Based on patch and suggestion provided by Dr. Luc Moulinier
<luc.moulinier@unistra.fr>.

Changes include:

- New optional -verbose switch to all Rtcl commands

- Updated man page and README

Signed-off-by: Matt Adams <matt.adams@radicaldynamic.com>